+++ Source/RSP/X86.h
@@ -46,7 +46,7 @@ void AdcConstToX86reg				( BYTE Constant, int x86reg );
 void AdcConstToVariable				( void *Variable, char *VariableName, BYTE Constant );
 void AdcConstHalfToVariable			( void *Variable, char *VariableName, BYTE Constant );
 void AddConstToVariable				( DWORD Const, void *Variable, char *VariableName );
-void AddConstToX86Reg				( int x86Reg, DWORD Const );
+void AddConstToX86Reg				( int x86Reg, size_t Const );
 void AddVariableToX86reg			( int x86reg, void * Variable, char * VariableName );
 void AddX86regToVariable			( int x86reg, void * Variable, char * VariableName );
 void AddX86regHalfToVariable		( int x86reg, void * Variable, char * VariableName );
